+++ baseapp/abci_test.go
@@ -1,4 +1,4 @@
-package baseapp_test
+package baseapp
 
 import (
 	"fmt"
@@ -9,7 +9,6 @@ import (
 	tmprototypes "github.com/tendermint/tendermint/proto/tendermint/types"
 	dbm "github.com/tendermint/tm-db"
 
-	"github.com/cosmos/cosmos-sdk/baseapp"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -19,81 +18,81 @@ func TestGetBlockRentionHeight(t *testing.T) {
 	name := t.Name()
 
 	testCases := map[string]struct {
-		bapp         *baseapp.BaseApp
+		bapp         *BaseApp
 		maxAgeBlocks int64
 		commitHeight int64
 		expected     int64
 	}{
 		"defaults": {
-			bapp:         baseapp.NewBaseApp(name, logger, db, nil),
+			bapp:         NewBaseApp(name, logger, db, nil),
 			maxAgeBlocks: 0,
 			commitHeight: 499000,
 			expected:     0,
 		},
 		"pruning unbonding time only": {
-			bapp:         baseapp.NewBaseApp(name, logger, db, nil, baseapp.SetMinRetainBlocks(1)),
+			bapp:         NewBaseApp(name, logger, db, nil, SetMinRetainBlocks(1)),
 			maxAgeBlocks: 362880,
 			commitHeight: 499000,
 			expected:     136120,
 		},
 		"pruning iavl snapshot only": {
-			bapp: baseapp.NewBaseApp(
+			bapp: NewBaseApp(
 				name, logger, db, nil,
-				baseapp.SetPruning(sdk.PruningOptions{KeepEvery: 10000}),
-				baseapp.SetMinRetainBlocks(1),
+				SetPruning(sdk.PruningOptions{KeepEvery: 10000}),
+				SetMinRetainBlocks(1),
 			),
 			maxAgeBlocks: 0,
 			commitHeight: 499000,
 			expected:     490000,
 		},
 		"pruning state sync snapshot only": {
-			bapp: baseapp.NewBaseApp(
+			bapp: NewBaseApp(
 				name, logger, db, nil,
-				baseapp.SetSnapshotInterval(50000),
-				baseapp.SetSnapshotKeepRecent(3),
-				baseapp.SetMinRetainBlocks(1),
+				SetSnapshotInterval(50000),
+				SetSnapshotKeepRecent(3),
+				SetMinRetainBlocks(1),
 			),
 			maxAgeBlocks: 0,
 			commitHeight: 499000,
 			expected:     349000,
 		},
 		"pruning min retention only": {
-			bapp: baseapp.NewBaseApp(
+			bapp: NewBaseApp(
 				name, logger, db, nil,
-				baseapp.SetMinRetainBlocks(400000),
+				SetMinRetainBlocks(400000),
 			),
 			maxAgeBlocks: 0,
 			commitHeight: 499000,
 			expected:     99000,
 		},
 		"pruning all conditions": {
-			bapp: baseapp.NewBaseApp(
+			bapp: NewBaseApp(
 				name, logger, db, nil,
-				baseapp.SetPruning(sdk.PruningOptions{KeepEvery: 10000}),
-				baseapp.SetMinRetainBlocks(400000),
-				baseapp.SetSnapshotInterval(50000), baseapp.SetSnapshotKeepRecent(3),
+				SetPruning(sdk.PruningOptions{KeepEvery: 10000}),
+				SetMinRetainBlocks(400000),
+				SetSnapshotInterval(50000), SetSnapshotKeepRecent(3),
 			),
 			maxAgeBlocks: 362880,
 			commitHeight: 499000,
 			expected:     99000,
 		},
 		"no pruning due to no persisted state": {
-			bapp: baseapp.NewBaseApp(
+			bapp: NewBaseApp(
 				name, logger, db, nil,
-				baseapp.SetPruning(sdk.PruningOptions{KeepEvery: 10000}),
-				baseapp.SetMinRetainBlocks(400000),
-				baseapp.SetSnapshotInterval(50000), baseapp.SetSnapshotKeepRecent(3),
+				SetPruning(sdk.PruningOptions{KeepEvery: 10000}),
+				SetMinRetainBlocks(400000),
+				SetSnapshotInterval(50000), SetSnapshotKeepRecent(3),
 			),
 			maxAgeBlocks: 362880,
 			commitHeight: 10000,
 			expected:     0,
 		},
 		"disable pruning": {
-			bapp: baseapp.NewBaseApp(
+			bapp: NewBaseApp(
 				name, logger, db, nil,
-				baseapp.SetPruning(sdk.PruningOptions{KeepEvery: 10000}),
-				baseapp.SetMinRetainBlocks(0),
-				baseapp.SetSnapshotInterval(50000), baseapp.SetSnapshotKeepRecent(3),
+				SetPruning(sdk.PruningOptions{KeepEvery: 10000}),
+				SetMinRetainBlocks(0),
+				SetSnapshotInterval(50000), SetSnapshotKeepRecent(3),
 			),
 			maxAgeBlocks: 362880,
 			commitHeight: 499000,
@@ -127,14 +126,14 @@ func TestBaseAppCreateQueryContextRejectsNegativeHeights(t *testing.T) {
 	logger := defaultLogger()
 	db := dbm.NewMemDB()
 	name := t.Name()
-	app := baseapp.NewBaseApp(name, logger, db, nil)
+	app := NewBaseApp(name, logger, db, nil)
 
 	proves := []bool{
 		false, true,
 	}
 	for _, prove := range proves {
 		t.Run(fmt.Sprintf("prove=%t", prove), func(t *testing.T) {
-			sctx, err := app.CreateQueryContext(-10, true)
+			sctx, err := app.createQueryContext(-10, true)
 			require.Error(t, err)
 			require.Equal(t, sctx, sdk.Context{})
 		})
