+++ baseapp/baseapp_test.go
@@ -1,4 +1,4 @@
-package baseapp_test
+package baseapp
 
 import (
 	"bytes"
@@ -8,12 +8,10 @@ import (
 	"math"
 	"math/rand"
 	"os"
-	"strings"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/gogo/protobuf/jsonpb"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	abci "github.com/tendermint/tendermint/abci/types"
@@ -21,7 +19,6 @@ import (
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	dbm "github.com/tendermint/tm-db"
 
-	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/snapshots"
 	snapshottypes "github.com/cosmos/cosmos-sdk/snapshots/types"
@@ -30,8 +27,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/types/tx"
-	"github.com/cosmos/cosmos-sdk/x/auth/middleware"
 	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 )
 
@@ -83,12 +78,12 @@ func defaultLogger() log.Logger {
 	return log.NewTMLogger(log.NewSyncWriter(os.Stdout)).With("module", "sdk/app")
 }
 
-func newBaseApp(name string, options ...func(*baseapp.BaseApp)) *baseapp.BaseApp {
+func newBaseApp(name string, options ...func(*BaseApp)) *BaseApp {
 	logger := defaultLogger()
 	db := dbm.NewMemDB()
 	codec := codec.NewLegacyAmino()
 	registerTestCodec(codec)
-	return baseapp.NewBaseApp(name, logger, db, testTxDecoder(codec), options...)
+	return NewBaseApp(name, logger, db, testTxDecoder(codec), options...)
 }
 
 func registerTestCodec(cdc *codec.LegacyAmino) {
@@ -112,7 +107,7 @@ func aminoTxEncoder() sdk.TxEncoder {
 }
 
 // simple one store baseapp
-func setupBaseApp(t *testing.T, options ...func(*baseapp.BaseApp)) *baseapp.BaseApp {
+func setupBaseApp(t *testing.T, options ...func(*BaseApp)) *BaseApp {
 	app := newBaseApp(t.Name(), options...)
 	require.Equal(t, t.Name(), app.Name())
 
@@ -125,39 +120,11 @@ func setupBaseApp(t *testing.T, options ...func(*baseapp.BaseApp)) *baseapp.Base
 	return app
 }
 
-// testTxHandler is a tx.Handler used for the mock app, it does not
-// contain any signature verification logic.
-func testTxHandler(options middleware.TxHandlerOptions, customTxHandlerMiddleware handlerFun) tx.Handler {
-	return middleware.ComposeMiddlewares(
-		middleware.NewRunMsgsTxHandler(options.MsgServiceRouter, options.LegacyRouter),
-		middleware.GasTxMiddleware,
-		middleware.RecoveryTxMiddleware,
-		middleware.NewIndexEventsTxMiddleware(options.IndexEvents),
-		middleware.ValidateBasicMiddleware,
-		CustomTxHandlerMiddleware(customTxHandlerMiddleware),
-	)
-}
-
 // simple one store baseapp with data and snapshots. Each tx is 1 MB in size (uncompressed).
-func setupBaseAppWithSnapshots(t *testing.T, blocks uint, blockTxs int, options ...func(*baseapp.BaseApp)) (*baseapp.BaseApp, func()) {
+func setupBaseAppWithSnapshots(t *testing.T, blocks uint, blockTxs int, options ...func(*BaseApp)) (*BaseApp, func()) {
 	codec := codec.NewLegacyAmino()
 	registerTestCodec(codec)
-	routerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		legacyRouter.AddRoute(sdk.NewRoute(routeMsgKeyValue, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-			kv := msg.(*msgKeyValue)
-			bapp.CMS().GetCommitKVStore(capKey2).Set(kv.Key, kv.Value)
-			return &sdk.Result{}, nil
-		}))
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			func(ctx sdk.Context, tx sdk.Tx, simulate bool) (sdk.Context, error) { return ctx, nil },
-		)
-		bapp.SetTxHandler(txHandler)
-	}
+	routerOpt := func(bapp *BaseApp) { bapp.SetTxHandler(nil) }
 
 	snapshotInterval := uint64(2)
 	snapshotTimeout := 1 * time.Minute
@@ -170,9 +137,9 @@ func setupBaseAppWithSnapshots(t *testing.T, blocks uint, blockTxs int, options 
 	}
 
 	app := setupBaseApp(t, append(options,
-		baseapp.SetSnapshotStore(snapshotStore),
-		baseapp.SetSnapshotInterval(snapshotInterval),
-		baseapp.SetPruning(sdk.PruningOptions{KeepEvery: 1}),
+		SetSnapshotStore(snapshotStore),
+		SetSnapshotInterval(snapshotInterval),
+		SetPruning(sdk.PruningOptions{KeepEvery: 1}),
 		routerOpt)...)
 
 	app.InitChain(abci.RequestInitChain{})
@@ -223,9 +190,9 @@ func TestMountStores(t *testing.T) {
 	app := setupBaseApp(t)
 
 	// check both stores
-	store1 := app.CMS().GetCommitKVStore(capKey1)
+	store1 := app.cms.GetCommitKVStore(capKey1)
 	require.NotNil(t, store1)
-	store2 := app.CMS().GetCommitKVStore(capKey2)
+	store2 := app.cms.GetCommitKVStore(capKey2)
 	require.NotNil(t, store2)
 }
 
@@ -233,10 +200,10 @@ func TestMountStores(t *testing.T) {
 // Test that LoadLatestVersion actually does.
 func TestLoadVersion(t *testing.T) {
 	logger := defaultLogger()
-	pruningOpt := baseapp.SetPruning(storetypes.PruneNothing)
+	pruningOpt := SetPruning(storetypes.PruneNothing)
 	db := dbm.NewMemDB()
 	name := t.Name()
-	app := baseapp.NewBaseApp(name, logger, db, nil, pruningOpt)
+	app := NewBaseApp(name, logger, db, nil, pruningOpt)
 
 	// make a cap key and mount the store
 	err := app.LoadLatestVersion() // needed to make stores non-nil
@@ -263,7 +230,7 @@ func TestLoadVersion(t *testing.T) {
 	commitID2 := storetypes.CommitID{Version: 2, Hash: res.Data}
 
 	// reload with LoadLatestVersion
-	app = baseapp.NewBaseApp(name, logger, db, nil, pruningOpt)
+	app = NewBaseApp(name, logger, db, nil, pruningOpt)
 	app.MountStores()
 	err = app.LoadLatestVersion()
 	require.Nil(t, err)
@@ -271,7 +238,7 @@ func TestLoadVersion(t *testing.T) {
 
 	// reload with LoadVersion, see if you can commit the same block and get
 	// the same result
-	app = baseapp.NewBaseApp(name, logger, db, nil, pruningOpt)
+	app = NewBaseApp(name, logger, db, nil, pruningOpt)
 	err = app.LoadVersion(1)
 	require.Nil(t, err)
 	testLoadVersionHelper(t, app, int64(1), commitID1)
@@ -280,8 +247,8 @@ func TestLoadVersion(t *testing.T) {
 	testLoadVersionHelper(t, app, int64(2), commitID2)
 }
 
-func useDefaultLoader(app *baseapp.BaseApp) {
-	app.SetStoreLoader(baseapp.DefaultStoreLoader)
+func useDefaultLoader(app *BaseApp) {
+	app.SetStoreLoader(DefaultStoreLoader)
 }
 
 func initStore(t *testing.T, db dbm.DB, storeKey string, k, v []byte) {
@@ -320,7 +287,7 @@ func checkStore(t *testing.T, db dbm.DB, ver int64, storeKey string, k, v []byte
 // Test that LoadLatestVersion actually does.
 func TestSetLoader(t *testing.T) {
 	cases := map[string]struct {
-		setLoader    func(*baseapp.BaseApp)
+		setLoader    func(*BaseApp)
 		origStoreKey string
 		loadStoreKey string
 	}{
@@ -346,11 +313,11 @@ func TestSetLoader(t *testing.T) {
 			initStore(t, db, tc.origStoreKey, k, v)
 
 			// load the app with the existing db
-			opts := []func(*baseapp.BaseApp){baseapp.SetPruning(storetypes.PruneNothing)}
+			opts := []func(*BaseApp){SetPruning(storetypes.PruneNothing)}
 			if tc.setLoader != nil {
 				opts = append(opts, tc.setLoader)
 			}
-			app := baseapp.NewBaseApp(t.Name(), defaultLogger(), db, nil, opts...)
+			app := NewBaseApp(t.Name(), defaultLogger(), db, nil, opts...)
 			app.MountStores(sdk.NewKVStoreKey(tc.loadStoreKey))
 			err := app.LoadLatestVersion()
 			require.Nil(t, err)
@@ -369,10 +336,10 @@ func TestSetLoader(t *testing.T) {
 
 func TestVersionSetterGetter(t *testing.T) {
 	logger := defaultLogger()
-	pruningOpt := baseapp.SetPruning(storetypes.PruneDefault)
+	pruningOpt := SetPruning(storetypes.PruneDefault)
 	db := dbm.NewMemDB()
 	name := t.Name()
-	app := baseapp.NewBaseApp(name, logger, db, nil, pruningOpt)
+	app := NewBaseApp(name, logger, db, nil, pruningOpt)
 
 	require.Equal(t, "", app.Version())
 	res := app.Query(abci.RequestQuery{Path: "app/version"})
@@ -389,10 +356,10 @@ func TestVersionSetterGetter(t *testing.T) {
 
 func TestLoadVersionInvalid(t *testing.T) {
 	logger := log.NewNopLogger()
-	pruningOpt := baseapp.SetPruning(storetypes.PruneNothing)
+	pruningOpt := SetPruning(storetypes.PruneNothing)
 	db := dbm.NewMemDB()
 	name := t.Name()
-	app := baseapp.NewBaseApp(name, logger, db, nil, pruningOpt)
+	app := NewBaseApp(name, logger, db, nil, pruningOpt)
 
 	err := app.LoadLatestVersion()
 	require.Nil(t, err)
@@ -407,7 +374,7 @@ func TestLoadVersionInvalid(t *testing.T) {
 	commitID1 := storetypes.CommitID{Version: 1, Hash: res.Data}
 
 	// create a new app with the stores mounted under the same cap key
-	app = baseapp.NewBaseApp(name, logger, db, nil, pruningOpt)
+	app = NewBaseApp(name, logger, db, nil, pruningOpt)
 
 	// require we can load the latest version
 	err = app.LoadVersion(1)
@@ -426,10 +393,10 @@ func TestLoadVersionPruning(t *testing.T) {
 		KeepEvery:  3,
 		Interval:   1,
 	}
-	pruningOpt := baseapp.SetPruning(pruningOptions)
+	pruningOpt := SetPruning(pruningOptions)
 	db := dbm.NewMemDB()
 	name := t.Name()
-	app := baseapp.NewBaseApp(name, logger, db, nil, pruningOpt)
+	app := NewBaseApp(name, logger, db, nil, pruningOpt)
 
 	// make a cap key and mount the store
 	capKey := sdk.NewKVStoreKey("key1")
@@ -457,17 +424,17 @@ func TestLoadVersionPruning(t *testing.T) {
 	}
 
 	for _, v := range []int64{1, 2, 4} {
-		_, err = app.CMS().CacheMultiStoreWithVersion(v)
+		_, err = app.cms.CacheMultiStoreWithVersion(v)
 		require.NoError(t, err)
 	}
 
 	for _, v := range []int64{3, 5, 6, 7} {
-		_, err = app.CMS().CacheMultiStoreWithVersion(v)
+		_, err = app.cms.CacheMultiStoreWithVersion(v)
 		require.NoError(t, err)
 	}
 
 	// reload with LoadLatestVersion, check it loads last version
-	app = baseapp.NewBaseApp(name, logger, db, nil, pruningOpt)
+	app = NewBaseApp(name, logger, db, nil, pruningOpt)
 	app.MountStores(capKey)
 
 	err = app.LoadLatestVersion()
@@ -475,7 +442,7 @@ func TestLoadVersionPruning(t *testing.T) {
 	testLoadVersionHelper(t, app, int64(7), lastCommitID)
 }
 
-func testLoadVersionHelper(t *testing.T, app *baseapp.BaseApp, expectedHeight int64, expectedID storetypes.CommitID) {
+func testLoadVersionHelper(t *testing.T, app *BaseApp, expectedHeight int64, expectedID storetypes.CommitID) {
 	lastHeight := app.LastBlockHeight()
 	lastID := app.LastCommitID()
 	require.Equal(t, expectedHeight, lastHeight)
@@ -485,12 +452,12 @@ func testLoadVersionHelper(t *testing.T, app *baseapp.BaseApp, expectedHeight in
 func TestOptionFunction(t *testing.T) {
 	logger := defaultLogger()
 	db := dbm.NewMemDB()
-	bap := baseapp.NewBaseApp("starting name", logger, db, nil, testChangeNameHelper("new name"))
-	require.Equal(t, bap.GetName(), "new name", "BaseApp should have had name changed via option function")
+	bap := NewBaseApp("starting name", logger, db, nil, testChangeNameHelper("new name"))
+	require.Equal(t, bap.name, "new name", "BaseApp should have had name changed via option function")
 }
 
-func testChangeNameHelper(name string) func(*baseapp.BaseApp) {
-	return func(bap *baseapp.BaseApp) {
+func testChangeNameHelper(name string) func(*BaseApp) {
+	return func(bap *BaseApp) {
 		bap.SetName(name)
 	}
 }
@@ -505,7 +472,7 @@ func TestTxDecoder(t *testing.T) {
 	tx := newTxCounter(1, 0)
 	txBytes := codec.MustMarshal(tx)
 
-	dTx, err := app.TxDecoder(txBytes)
+	dTx, err := app.txDecoder(txBytes)
 	require.NoError(t, err)
 
 	cTx := dTx.(txTest)
@@ -570,8 +537,8 @@ func TestBaseAppOptionSeal(t *testing.T) {
 
 func TestSetMinGasPrices(t *testing.T) {
 	minGasPrices := sdk.DecCoins{sdk.NewInt64DecCoin("stake", 5000)}
-	app := newBaseApp(t.Name(), baseapp.SetMinGasPrices(minGasPrices.String()))
-	require.Equal(t, minGasPrices, app.MinGasPrices())
+	app := newBaseApp(t.Name(), SetMinGasPrices(minGasPrices.String()))
+	require.Equal(t, minGasPrices, app.minGasPrices)
 }
 
 func TestInitChainer(t *testing.T) {
@@ -580,7 +547,7 @@ func TestInitChainer(t *testing.T) {
 	// we can reload the same  app later
 	db := dbm.NewMemDB()
 	logger := defaultLogger()
-	app := baseapp.NewBaseApp(name, logger, db, nil)
+	app := NewBaseApp(name, logger, db, nil)
 	capKey := sdk.NewKVStoreKey("main")
 	capKey2 := sdk.NewKVStoreKey("key2")
 	app.MountStores(capKey, capKey2)
@@ -623,10 +590,10 @@ func TestInitChainer(t *testing.T) {
 	)
 
 	// assert that chainID is set correctly in InitChain
-	chainID := app.DeliverState().Context().ChainID()
+	chainID := app.deliverState.Context().ChainID()
 	require.Equal(t, "test-chain-id", chainID, "ChainID in deliverState not set correctly in InitChain")
 
-	chainID = app.CheckState().Context().ChainID()
+	chainID = app.checkState.Context().ChainID()
 	require.Equal(t, "test-chain-id", chainID, "ChainID in checkState not set correctly in InitChain")
 
 	app.Commit()
@@ -635,7 +602,7 @@ func TestInitChainer(t *testing.T) {
 	require.Equal(t, value, res.Value)
 
 	// reload app
-	app = baseapp.NewBaseApp(name, logger, db, nil)
+	app = NewBaseApp(name, logger, db, nil)
 	app.SetInitChainer(initChainer)
 	app.MountStores(capKey, capKey2)
 	err = app.LoadLatestVersion() // needed to make stores non-nil
@@ -659,7 +626,7 @@ func TestInitChain_WithInitialHeight(t *testing.T) {
 	name := t.Name()
 	db := dbm.NewMemDB()
 	logger := defaultLogger()
-	app := baseapp.NewBaseApp(name, logger, db, nil)
+	app := NewBaseApp(name, logger, db, nil)
 
 	app.InitChain(
 		abci.RequestInitChain{
@@ -675,7 +642,7 @@ func TestBeginBlock_WithInitialHeight(t *testing.T) {
 	name := t.Name()
 	db := dbm.NewMemDB()
 	logger := defaultLogger()
-	app := baseapp.NewBaseApp(name, logger, db, nil)
+	app := NewBaseApp(name, logger, db, nil)
 
 	app.InitChain(
 		abci.RequestInitChain{
@@ -844,68 +811,6 @@ func testTxDecoder(cdc *codec.LegacyAmino) sdk.TxDecoder {
 	}
 }
 
-func customHandlerTxTest(t *testing.T, capKey storetypes.StoreKey, storeKey []byte) handlerFun {
-	return func(ctx sdk.Context, tx sdk.Tx, simulate bool) (sdk.Context, error) {
-		store := ctx.KVStore(capKey)
-		txTest := tx.(txTest)
-
-		if txTest.FailOnAnte {
-			return ctx, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "ante handler failure")
-		}
-
-		_, err := incrementingCounter(t, store, storeKey, txTest.Counter)
-		if err != nil {
-			return ctx, err
-		}
-
-		ctx.EventManager().EmitEvents(
-			counterEvent("post_handlers", txTest.Counter),
-		)
-
-		return ctx, nil
-	}
-}
-
-func counterEvent(evType string, msgCount int64) sdk.Events {
-	return sdk.Events{
-		sdk.NewEvent(
-			evType,
-			sdk.NewAttribute("update_counter", fmt.Sprintf("%d", msgCount)),
-		),
-	}
-}
-
-func handlerMsgCounter(t *testing.T, capKey storetypes.StoreKey, deliverKey []byte) sdk.Handler {
-	return func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		ctx = ctx.WithEventManager(sdk.NewEventManager())
-		store := ctx.KVStore(capKey)
-		var msgCount int64
-
-		switch m := msg.(type) {
-		case *msgCounter:
-			if m.FailOnHandler {
-				return nil, sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "message handler failure")
-			}
-
-			msgCount = m.Counter
-		case *msgCounter2:
-			msgCount = m.Counter
-		}
-
-		ctx.EventManager().EmitEvents(
-			counterEvent(sdk.EventTypeMessage, msgCount),
-		)
-
-		res, err := incrementingCounter(t, store, deliverKey, msgCount)
-		if err != nil {
-			return nil, err
-		}
-
-		res.Events = ctx.EventManager().Events().ToABCIEvents()
-		return res, nil
-	}
-}
-
 func getIntFromStore(store sdk.KVStore, key []byte) int64 {
 	bz := store.Get(key)
 	if len(bz) == 0 {
@@ -933,781 +838,73 @@ func incrementingCounter(t *testing.T, store sdk.KVStore, counterKey []byte, cou
 	return &sdk.Result{}, nil
 }
 
-// ---------------------------------------------------------------------
-// Tx processing - CheckTx, DeliverTx, SimulateTx.
-// These tests use the serialized tx as input, while most others will use the
-// Check(), Deliver(), Simulate() methods directly.
-// Ensure that Check/Deliver/Simulate work as expected with the store.
-
-// Test that successive CheckTx can see each others' effects
-// on the store within a block, and that the CheckTx state
-// gets reset to the latest committed state during Commit
-func TestCheckTx(t *testing.T) {
-	// This ante handler reads the key and checks that the value matches the current counter.
-	// This ensures changes to the kvstore persist across successive CheckTx.
-	counterKey := []byte("counter-key")
-
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		// TODO: can remove this once CheckTx doesnt process msgs.
-		legacyRouter.AddRoute(sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-			return &sdk.Result{}, nil
-		}))
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			customHandlerTxTest(t, capKey1, counterKey),
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-
-	app := setupBaseApp(t, txHandlerOpt)
-
-	nTxs := int64(5)
-	app.InitChain(abci.RequestInitChain{})
-
-	// Create same codec used in txDecoder
-	codec := codec.NewLegacyAmino()
-	registerTestCodec(codec)
-
-	for i := int64(0); i < nTxs; i++ {
-		tx := newTxCounter(i, 0) // no messages
-		txBytes, err := codec.Marshal(tx)
-		require.NoError(t, err)
-		r := app.CheckTx(abci.RequestCheckTx{Tx: txBytes})
-		require.Empty(t, r.GetEvents())
-		require.True(t, r.IsOK(), fmt.Sprintf("%v", r))
-	}
-
-	checkStateStore := app.CheckState().Context().KVStore(capKey1)
-	storedCounter := getIntFromStore(checkStateStore, counterKey)
-
-	// Ensure storedCounter
-	require.Equal(t, nTxs, storedCounter)
-
-	// If a block is committed, CheckTx state should be reset.
-	header := tmproto.Header{Height: 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header, Hash: []byte("hash")})
-
-	require.NotNil(t, app.CheckState().Context().BlockGasMeter(), "block gas meter should have been set to checkState")
-	require.NotEmpty(t, app.CheckState().Context().HeaderHash())
-
-	app.EndBlock(abci.RequestEndBlock{})
-	app.Commit()
-
-	checkStateStore = app.CheckState().Context().KVStore(capKey1)
-	storedBytes := checkStateStore.Get(counterKey)
-	require.Nil(t, storedBytes)
-}
-
-// Test that successive DeliverTx can see each others' effects
-// on the store, both within and across blocks.
-func TestDeliverTx(t *testing.T) {
-	// test increments in the post txHandler
-	anteKey := []byte("ante-key")
-	// test increments in the handler
-	deliverKey := []byte("deliver-key")
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r := sdk.NewRoute(routeMsgCounter, handlerMsgCounter(t, capKey1, deliverKey))
-		legacyRouter.AddRoute(r)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			customHandlerTxTest(t, capKey1, anteKey),
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-	app.InitChain(abci.RequestInitChain{})
-
-	// Create same codec used in txDecoder
-	codec := codec.NewLegacyAmino()
-	registerTestCodec(codec)
-
-	nBlocks := 3
-	txPerHeight := 5
-
-	for blockN := 0; blockN < nBlocks; blockN++ {
-		header := tmproto.Header{Height: int64(blockN) + 1}
-		app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-		for i := 0; i < txPerHeight; i++ {
-			counter := int64(blockN*txPerHeight + i)
-			tx := newTxCounter(counter, counter)
-
-			txBytes, err := codec.Marshal(tx)
-			require.NoError(t, err)
-
-			res := app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-			require.True(t, res.IsOK(), fmt.Sprintf("%v", res))
-			events := res.GetEvents()
-			require.Len(t, events, 3, "should contain ante handler, message type and counter events respectively")
-			require.Equal(t, sdk.MarkEventsToIndex(counterEvent("post_handlers", counter).ToABCIEvents(), map[string]struct{}{})[0], events[0], "ante handler event")
-			require.Equal(t, sdk.MarkEventsToIndex(counterEvent(sdk.EventTypeMessage, counter).ToABCIEvents(), map[string]struct{}{})[0], events[2], "msg handler update counter event")
-		}
-
-		app.EndBlock(abci.RequestEndBlock{})
-		app.Commit()
-	}
-}
-
-// Number of messages doesn't matter to CheckTx.
-func TestMultiMsgCheckTx(t *testing.T) {
-	// TODO: ensure we get the same results
-	// with one message or many
-}
-
-// One call to DeliverTx should process all the messages, in order.
-func TestMultiMsgDeliverTx(t *testing.T) {
-	// increment the tx counter
-	anteKey := []byte("ante-key")
-	// increment the msg counter
-	deliverKey := []byte("deliver-key")
-	deliverKey2 := []byte("deliver-key2")
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r1 := sdk.NewRoute(routeMsgCounter, handlerMsgCounter(t, capKey1, deliverKey))
-		r2 := sdk.NewRoute(routeMsgCounter2, handlerMsgCounter(t, capKey1, deliverKey2))
-		legacyRouter.AddRoute(r1)
-		legacyRouter.AddRoute(r2)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			customHandlerTxTest(t, capKey1, anteKey),
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-
-	// Create same codec used in txDecoder
-	codec := codec.NewLegacyAmino()
-	registerTestCodec(codec)
-
-	// run a multi-msg tx
-	// with all msgs the same route
-
-	header := tmproto.Header{Height: 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	tx := newTxCounter(0, 0, 1, 2)
-	txBytes, err := codec.Marshal(tx)
-	require.NoError(t, err)
-	res := app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-	require.True(t, res.IsOK(), fmt.Sprintf("%v", res))
-
-	store := app.DeliverState().Context().KVStore(capKey1)
-
-	// tx counter only incremented once
-	txCounter := getIntFromStore(store, anteKey)
-	require.Equal(t, int64(1), txCounter)
-
-	// msg counter incremented three times
-	msgCounter := getIntFromStore(store, deliverKey)
-	require.Equal(t, int64(3), msgCounter)
-
-	// replace the second message with a msgCounter2
-
-	tx = newTxCounter(1, 3)
-	tx.Msgs = append(tx.Msgs, msgCounter2{0})
-	tx.Msgs = append(tx.Msgs, msgCounter2{1})
-	txBytes, err = codec.Marshal(tx)
-	require.NoError(t, err)
-	res = app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-	require.True(t, res.IsOK(), fmt.Sprintf("%v", res))
-
-	store = app.DeliverState().Context().KVStore(capKey1)
-
-	// tx counter only incremented once
-	txCounter = getIntFromStore(store, anteKey)
-	require.Equal(t, int64(2), txCounter)
-
-	// original counter increments by one
-	// new counter increments by two
-	msgCounter = getIntFromStore(store, deliverKey)
-	require.Equal(t, int64(4), msgCounter)
-	msgCounter2 := getIntFromStore(store, deliverKey2)
-	require.Equal(t, int64(2), msgCounter2)
-}
-
-// Interleave calls to Check and Deliver and ensure
-// that there is no cross-talk. Check sees results of the previous Check calls
-// and Deliver sees that of the previous Deliver calls, but they don't see eachother.
-func TestConcurrentCheckDeliver(t *testing.T) {
-	// TODO
-}
-
-// Simulate a transaction that uses gas to compute the gas.
-// Simulate() and Query("/app/simulate", txBytes) should give
-// the same results.
-func TestSimulateTx(t *testing.T) {
-	gasConsumed := uint64(5)
-
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-			ctx.GasMeter().ConsumeGas(gasConsumed, "test")
-			return &sdk.Result{}, nil
-		})
-		legacyRouter.AddRoute(r)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			func(ctx sdk.Context, tx sdk.Tx, simulate bool) (sdk.Context, error) { return ctx, nil },
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-
-	app.InitChain(abci.RequestInitChain{})
-
-	// Create same codec used in txDecoder
-	cdc := codec.NewLegacyAmino()
-	registerTestCodec(cdc)
-
-	nBlocks := 3
-	for blockN := 0; blockN < nBlocks; blockN++ {
-		count := int64(blockN + 1)
-		header := tmproto.Header{Height: count}
-		app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-		tx := newTxCounter(count, count)
-		tx.GasLimit = gasConsumed
-		txBytes, err := cdc.Marshal(tx)
-		require.Nil(t, err)
-
-		// simulate a message, check gas reported
-		gInfo, result, err := app.Simulate(txBytes)
-		require.NoError(t, err)
-		require.NotNil(t, result)
-		require.Equal(t, gasConsumed, gInfo.GasUsed)
-
-		// simulate again, same result
-		gInfo, result, err = app.Simulate(txBytes)
-		require.NoError(t, err)
-		require.NotNil(t, result)
-		require.Equal(t, gasConsumed, gInfo.GasUsed)
-
-		// simulate by calling Query with encoded tx
-		query := abci.RequestQuery{
-			Path: "/app/simulate",
-			Data: txBytes,
-		}
-		queryResult := app.Query(query)
-		require.True(t, queryResult.IsOK(), queryResult.Log)
-
-		var simRes sdk.SimulationResponse
-		require.NoError(t, jsonpb.Unmarshal(strings.NewReader(string(queryResult.Value)), &simRes))
-
-		require.Equal(t, gInfo, simRes.GasInfo)
-		require.Equal(t, result.Log, simRes.Result.Log)
-		require.Equal(t, result.Events, simRes.Result.Events)
-		require.True(t, bytes.Equal(result.Data, simRes.Result.Data))
-
-		app.EndBlock(abci.RequestEndBlock{})
-		app.Commit()
-	}
-}
-
-func TestRunInvalidTransaction(t *testing.T) {
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-			return &sdk.Result{}, nil
-		})
-		legacyRouter.AddRoute(r)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			func(ctx sdk.Context, tx sdk.Tx, simulate bool) (newCtx sdk.Context, err error) {
-				return
-			},
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-
-	header := tmproto.Header{Height: 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-	// transaction with no messages
-	{
-		emptyTx := &txTest{}
-		_, result, err := app.SimDeliver(aminoTxEncoder(), emptyTx)
-		require.Nil(t, result)
-
-		space, code, _ := sdkerrors.ABCIInfo(err, false)
-		require.EqualValues(t, sdkerrors.ErrInvalidRequest.Codespace(), space, err)
-		require.EqualValues(t, sdkerrors.ErrInvalidRequest.ABCICode(), code, err)
-	}
-
-	// transaction where ValidateBasic fails
-	{
-		testCases := []struct {
-			tx   txTest
-			fail bool
-		}{
-			{newTxCounter(0, 0), false},
-			{newTxCounter(-1, 0), false},
-			{newTxCounter(100, 100), false},
-			{newTxCounter(100, 5, 4, 3, 2, 1), false},
-
-			{newTxCounter(0, -1), true},
-			{newTxCounter(0, 1, -2), true},
-			{newTxCounter(0, 1, 2, -10, 5), true},
-		}
-
-		for _, testCase := range testCases {
-			tx := testCase.tx
-			_, result, err := app.SimDeliver(aminoTxEncoder(), tx)
-
-			if testCase.fail {
-				require.Error(t, err)
-
-				space, code, _ := sdkerrors.ABCIInfo(err, false)
-				require.EqualValues(t, sdkerrors.ErrInvalidSequence.Codespace(), space, err)
-				require.EqualValues(t, sdkerrors.ErrInvalidSequence.ABCICode(), code, err)
-			} else {
-				require.NotNil(t, result)
-			}
-		}
-	}
-
-	// transaction with no known route
-	{
-		unknownRouteTx := txTest{[]sdk.Msg{msgNoRoute{}}, 0, false, math.MaxUint64}
-		_, result, err := app.SimDeliver(aminoTxEncoder(), unknownRouteTx)
-		require.Error(t, err)
-		require.Nil(t, result)
-
-		space, code, _ := sdkerrors.ABCIInfo(err, false)
-		require.EqualValues(t, sdkerrors.ErrUnknownRequest.Codespace(), space, err)
-		require.EqualValues(t, sdkerrors.ErrUnknownRequest.ABCICode(), code, err)
-
-		unknownRouteTx = txTest{[]sdk.Msg{msgCounter{}, msgNoRoute{}}, 0, false, math.MaxUint64}
-		_, result, err = app.SimDeliver(aminoTxEncoder(), unknownRouteTx)
-		require.Error(t, err)
-		require.Nil(t, result)
-
-		space, code, _ = sdkerrors.ABCIInfo(err, false)
-		require.EqualValues(t, sdkerrors.ErrUnknownRequest.Codespace(), space, err)
-		require.EqualValues(t, sdkerrors.ErrUnknownRequest.ABCICode(), code, err)
-	}
-
-	// Transaction with an unregistered message
-	{
-		tx := newTxCounter(0, 0)
-		tx.Msgs = append(tx.Msgs, msgNoDecode{})
-
-		// new codec so we can encode the tx, but we shouldn't be able to decode
-		newCdc := codec.NewLegacyAmino()
-		registerTestCodec(newCdc)
-		newCdc.RegisterConcrete(&msgNoDecode{}, "cosmos-sdk/baseapp/msgNoDecode", nil)
-
-		txBytes, err := newCdc.Marshal(tx)
-		require.NoError(t, err)
-
-		res := app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-		require.EqualValues(t, sdkerrors.ErrTxDecode.ABCICode(), res.Code)
-		require.EqualValues(t, sdkerrors.ErrTxDecode.Codespace(), res.Codespace)
-	}
-}
-
-// Test that transactions exceeding gas limits fail
-func TestTxGasLimits(t *testing.T) {
-	gasGranted := uint64(10)
-
-	ante := func(ctx sdk.Context, tx sdk.Tx, simulate bool) (sdk.Context, error) {
-		count := tx.(txTest).Counter
-		ctx.GasMeter().ConsumeGas(uint64(count), "counter-ante")
-		return ctx, nil
-	}
-
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-			count := msg.(msgCounter).Counter
-			ctx.GasMeter().ConsumeGas(uint64(count), "counter-handler")
-			return &sdk.Result{}, nil
-		})
-		legacyRouter.AddRoute(r)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			ante,
-		)
-
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-
-	header := tmproto.Header{Height: 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-	testCases := []struct {
-		tx      txTest
-		gasUsed uint64
-		fail    bool
-	}{
-		{newTxCounter(0, 0), 0, false},
-		{newTxCounter(1, 1), 2, false},
-		{newTxCounter(9, 1), 10, false},
-		{newTxCounter(1, 9), 10, false},
-		{newTxCounter(10, 0), 10, false},
-		{newTxCounter(0, 10), 10, false},
-		{newTxCounter(0, 8, 2), 10, false},
-		{newTxCounter(0, 5, 1, 1, 1, 1, 1), 10, false},
-		{newTxCounter(0, 5, 1, 1, 1, 1), 9, false},
-
-		{newTxCounter(9, 2), 11, true},
-		{newTxCounter(2, 9), 11, true},
-		{newTxCounter(9, 1, 1), 11, true},
-		{newTxCounter(1, 8, 1, 1), 11, true},
-		{newTxCounter(11, 0), 11, true},
-		{newTxCounter(0, 11), 11, true},
-		{newTxCounter(0, 5, 11), 16, true},
-	}
-
-	for i, tc := range testCases {
-		tx := tc.tx
-		tx.GasLimit = gasGranted
-		gInfo, result, err := app.SimDeliver(aminoTxEncoder(), tx)
-
-		// check gas used and wanted
-		require.Equal(t, tc.gasUsed, gInfo.GasUsed, fmt.Sprintf("tc #%d; gas: %v, result: %v, err: %s", i, gInfo, result, err))
-
-		// check for out of gas
-		if !tc.fail {
-			require.NotNil(t, result, fmt.Sprintf("%d: %v, %v", i, tc, err))
-		} else {
-			require.Error(t, err)
-			require.Nil(t, result)
-
-			space, code, _ := sdkerrors.ABCIInfo(err, false)
-			require.EqualValues(t, sdkerrors.ErrOutOfGas.Codespace(), space, err)
-			require.EqualValues(t, sdkerrors.ErrOutOfGas.ABCICode(), code, err)
-		}
-	}
-}
-
-// Test that transactions exceeding gas limits fail
-func TestMaxBlockGasLimits(t *testing.T) {
-	gasGranted := uint64(10)
-	ante := func(ctx sdk.Context, tx sdk.Tx, simulate bool) (sdk.Context, error) {
-		count := tx.(txTest).Counter
-		ctx.GasMeter().ConsumeGas(uint64(count), "counter-ante")
-
-		return ctx, nil
-	}
-
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-			count := msg.(msgCounter).Counter
-			ctx.GasMeter().ConsumeGas(uint64(count), "counter-handler")
-			return &sdk.Result{}, nil
-		})
-		legacyRouter.AddRoute(r)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			ante,
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-
-	app.InitChain(abci.RequestInitChain{
-		ConsensusParams: &abci.ConsensusParams{
-			Block: &abci.BlockParams{
-				MaxGas: 100,
-			},
-		},
-	})
-
-	testCases := []struct {
-		tx                txTest
-		numDelivers       int
-		gasUsedPerDeliver uint64
-		fail              bool
-		failAfterDeliver  int
-	}{
-		{newTxCounter(0, 0), 0, 0, false, 0},
-		{newTxCounter(9, 1), 2, 10, false, 0},
-		{newTxCounter(10, 0), 3, 10, false, 0},
-		{newTxCounter(10, 0), 10, 10, false, 0},
-		{newTxCounter(2, 7), 11, 9, false, 0},
-		{newTxCounter(10, 0), 10, 10, false, 0}, // hit the limit but pass
-
-		{newTxCounter(10, 0), 11, 10, true, 10},
-		{newTxCounter(10, 0), 15, 10, true, 10},
-		{newTxCounter(9, 0), 12, 9, true, 11}, // fly past the limit
-	}
-
-	for i, tc := range testCases {
-		tx := tc.tx
-		tx.GasLimit = gasGranted
-
-		// reset the block gas
-		header := tmproto.Header{Height: app.LastBlockHeight() + 1}
-		app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-		// execute the transaction multiple times
-		for j := 0; j < tc.numDelivers; j++ {
-			_, result, err := app.SimDeliver(aminoTxEncoder(), tx)
-
-			ctx := app.DeliverState().Context()
-
-			// check for failed transactions
-			if tc.fail && (j+1) > tc.failAfterDeliver {
-				require.Error(t, err, fmt.Sprintf("tc #%d; result: %v, err: %s", i, result, err))
-				require.Nil(t, result, fmt.Sprintf("tc #%d; result: %v, err: %s", i, result, err))
-
-				space, code, _ := sdkerrors.ABCIInfo(err, false)
-				require.EqualValues(t, sdkerrors.ErrOutOfGas.Codespace(), space, err)
-				require.EqualValues(t, sdkerrors.ErrOutOfGas.ABCICode(), code, err)
-				require.True(t, ctx.BlockGasMeter().IsOutOfGas())
-			} else {
-				// check gas used and wanted
-				blockGasUsed := ctx.BlockGasMeter().GasConsumed()
-				expBlockGasUsed := tc.gasUsedPerDeliver * uint64(j+1)
-				require.Equal(
-					t, expBlockGasUsed, blockGasUsed,
-					fmt.Sprintf("%d,%d: %v, %v, %v, %v", i, j, tc, expBlockGasUsed, blockGasUsed, result),
-				)
-
-				require.NotNil(t, result, fmt.Sprintf("tc #%d; currDeliver: %d, result: %v, err: %s", i, j, result, err))
-				require.False(t, ctx.BlockGasMeter().IsPastLimit())
-			}
-		}
-	}
-}
-
-func TestBaseAppMiddleware(t *testing.T) {
-	anteKey := []byte("ante-key")
-	deliverKey := []byte("deliver-key")
-	cdc := codec.NewLegacyAmino()
-
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r := sdk.NewRoute(routeMsgCounter, handlerMsgCounter(t, capKey1, deliverKey))
-		legacyRouter.AddRoute(r)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			customHandlerTxTest(t, capKey1, anteKey),
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-
-	app.InitChain(abci.RequestInitChain{})
-	registerTestCodec(cdc)
-
-	header := tmproto.Header{Height: app.LastBlockHeight() + 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-	// execute a tx that will fail ante handler execution
-	//
-	// NOTE: State should not be mutated here. This will be implicitly checked by
-	// the next txs ante handler execution (customHandlerTxTest).
-	tx := newTxCounter(0, 0)
-	tx.setFailOnAnte(true)
-	txBytes, err := cdc.Marshal(tx)
-	require.NoError(t, err)
-	res := app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-	require.Empty(t, res.Events)
-	require.False(t, res.IsOK(), fmt.Sprintf("%v", res))
-
-	ctx := app.DeliverState().Context()
-	store := ctx.KVStore(capKey1)
-	require.Equal(t, int64(0), getIntFromStore(store, anteKey))
-
-	// execute at tx that will pass the ante handler (the checkTx state should
-	// mutate) but will fail the message handler
-	tx = newTxCounter(0, 0)
-	tx.setFailOnHandler(true)
-
-	txBytes, err = cdc.Marshal(tx)
-	require.NoError(t, err)
-
-	res = app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-	require.Empty(t, res.Events)
-	require.False(t, res.IsOK(), fmt.Sprintf("%v", res))
-
-	ctx = app.DeliverState().Context()
-	store = ctx.KVStore(capKey1)
-	require.Equal(t, int64(1), getIntFromStore(store, anteKey))
-	require.Equal(t, int64(0), getIntFromStore(store, deliverKey))
-
-	// execute a successful ante handler and message execution where state is
-	// implicitly checked by previous tx executions
-	tx = newTxCounter(1, 0)
-
-	txBytes, err = cdc.Marshal(tx)
-	require.NoError(t, err)
-
-	res = app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-	require.NotEmpty(t, res.Events)
-	require.True(t, res.IsOK(), fmt.Sprintf("%v", res))
-
-	ctx = app.DeliverState().Context()
-	store = ctx.KVStore(capKey1)
-	require.Equal(t, int64(2), getIntFromStore(store, anteKey))
-	require.Equal(t, int64(1), getIntFromStore(store, deliverKey))
-
-	// commit
-	app.EndBlock(abci.RequestEndBlock{})
-	app.Commit()
-}
-
-func TestGasConsumptionBadTx(t *testing.T) {
-	gasWanted := uint64(5)
-	ante := func(ctx sdk.Context, tx sdk.Tx, simulate bool) (sdk.Context, error) {
-		txTest := tx.(txTest)
-		ctx.GasMeter().ConsumeGas(uint64(txTest.Counter), "counter-ante")
-		if txTest.FailOnAnte {
-			return ctx, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "ante handler failure")
-		}
-
-		return ctx, nil
-	}
-
-	cdc := codec.NewLegacyAmino()
-	registerTestCodec(cdc)
-
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-			count := msg.(msgCounter).Counter
-			ctx.GasMeter().ConsumeGas(uint64(count), "counter-handler")
-			return &sdk.Result{}, nil
-		})
-		legacyRouter.AddRoute(r)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			ante,
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-
-	app.InitChain(abci.RequestInitChain{
-		ConsensusParams: &abci.ConsensusParams{
-			Block: &abci.BlockParams{
-				MaxGas: 9,
-			},
-		},
-	})
-
-	app.InitChain(abci.RequestInitChain{})
-
-	header := tmproto.Header{Height: app.LastBlockHeight() + 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-	tx := newTxCounter(5, 0)
-	tx.GasLimit = gasWanted
-	tx.setFailOnAnte(true)
-	txBytes, err := cdc.Marshal(tx)
-	require.NoError(t, err)
-
-	res := app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-	require.False(t, res.IsOK(), fmt.Sprintf("%v", res))
-
-	// require next tx to fail due to black gas limit
-	tx = newTxCounter(5, 0)
-	txBytes, err = cdc.Marshal(tx)
-	require.NoError(t, err)
-
-	res = app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-	require.False(t, res.IsOK(), fmt.Sprintf("%v", res))
-}
-
 // Test that we can only query from the latest committed state.
-func TestQuery(t *testing.T) {
-	key, value := []byte("hello"), []byte("goodbye")
+// func TestQuery(t *testing.T) {
+// 	key, value := []byte("hello"), []byte("goodbye")
 
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		legacyRouter := middleware.NewLegacyRouter()
-		r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-			store := ctx.KVStore(capKey1)
-			store.Set(key, value)
-			return &sdk.Result{}, nil
-		})
-		legacyRouter.AddRoute(r)
-		txHandler := testTxHandler(
-			middleware.TxHandlerOptions{
-				LegacyRouter:     legacyRouter,
-				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
-			},
-			func(ctx sdk.Context, tx sdk.Tx, simulate bool) (newCtx sdk.Context, err error) {
-				store := ctx.KVStore(capKey1)
-				store.Set(key, value)
-				return
-			},
-		)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
+// 	txHandlerOpt := func(bapp *BaseApp) {
+// 		legacyRouter := middleware.NewLegacyRouter()
+// 		r := sdk.NewRoute(routeMsgCounter, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
+// 			store := ctx.KVStore(capKey1)
+// 			store.Set(key, value)
+// 			return &sdk.Result{}, nil
+// 		})
+// 		legacyRouter.AddRoute(r)
+// 		txHandler := testTxHandler(
+// 			middleware.TxHandlerOptions{
+// 				LegacyRouter:     legacyRouter,
+// 				MsgServiceRouter: middleware.NewMsgServiceRouter(interfaceRegistry),
+// 			},
+// 			func(ctx sdk.Context, tx sdk.Tx, simulate bool) (newCtx sdk.Context, err error) {
+// 				store := ctx.KVStore(capKey1)
+// 				store.Set(key, value)
+// 				return
+// 			},
+// 		)
+// 		bapp.SetTxHandler(txHandler)
+// 	}
+// 	app := setupBaseApp(t, txHandlerOpt)
 
-	app.InitChain(abci.RequestInitChain{})
+// 	app.InitChain(abci.RequestInitChain{})
 
-	// NOTE: "/store/key1" tells us KVStore
-	// and the final "/key" says to use the data as the
-	// key in the given KVStore ...
-	query := abci.RequestQuery{
-		Path: "/store/key1/key",
-		Data: key,
-	}
-	tx := newTxCounter(0, 0)
+// 	// NOTE: "/store/key1" tells us KVStore
+// 	// and the final "/key" says to use the data as the
+// 	// key in the given KVStore ...
+// 	query := abci.RequestQuery{
+// 		Path: "/store/key1/key",
+// 		Data: key,
+// 	}
+// 	tx := newTxCounter(0, 0)
 
-	// query is empty before we do anything
-	res := app.Query(query)
-	require.Equal(t, 0, len(res.Value))
+// 	// query is empty before we do anything
+// 	res := app.Query(query)
+// 	require.Equal(t, 0, len(res.Value))
 
-	// query is still empty after a CheckTx
-	_, resTx, err := app.SimCheck(aminoTxEncoder(), tx)
-	require.NoError(t, err)
-	require.NotNil(t, resTx)
-	res = app.Query(query)
-	require.Equal(t, 0, len(res.Value))
+// 	// query is still empty after a CheckTx
+// 	_, resTx, err := app.SimCheck(aminoTxEncoder(), tx)
+// 	require.NoError(t, err)
+// 	require.NotNil(t, resTx)
+// 	res = app.Query(query)
+// 	require.Equal(t, 0, len(res.Value))
 
-	// query is still empty after a DeliverTx before we commit
-	header := tmproto.Header{Height: app.LastBlockHeight() + 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
+// 	// query is still empty after a DeliverTx before we commit
+// 	header := tmproto.Header{Height: app.LastBlockHeight() + 1}
+// 	app.BeginBlock(abci.RequestBeginBlock{Header: header})
 
-	_, resTx, err = app.SimDeliver(aminoTxEncoder(), tx)
-	require.NoError(t, err)
-	require.NotNil(t, resTx)
-	res = app.Query(query)
-	require.Equal(t, 0, len(res.Value))
+// 	_, resTx, err = app.SimDeliver(aminoTxEncoder(), tx)
+// 	require.NoError(t, err)
+// 	require.NotNil(t, resTx)
+// 	res = app.Query(query)
+// 	require.Equal(t, 0, len(res.Value))
 
-	// query returns correct value after Commit
-	app.Commit()
-	res = app.Query(query)
-	require.Equal(t, value, res.Value)
-}
+// 	// query returns correct value after Commit
+// 	app.Commit()
+// 	res = app.Query(query)
+// 	require.Equal(t, value, res.Value)
+// }
 
 func TestGRPCQuery(t *testing.T) {
-	grpcQueryOpt := func(bapp *baseapp.BaseApp) {
+	grpcQueryOpt := func(bapp *BaseApp) {
 		testdata.RegisterQueryServer(
 			bapp.GRPCQueryRouter(),
 			testdata.QueryImpl{},
@@ -1742,14 +939,14 @@ func TestGRPCQuery(t *testing.T) {
 
 // Test p2p filter queries
 func TestP2PQuery(t *testing.T) {
-	addrPeerFilterOpt := func(bapp *baseapp.BaseApp) {
+	addrPeerFilterOpt := func(bapp *BaseApp) {
 		bapp.SetAddrPeerFilter(func(addrport string) abci.ResponseQuery {
 			require.Equal(t, "1.1.1.1:8000", addrport)
 			return abci.ResponseQuery{Code: uint32(3)}
 		})
 	}
 
-	idPeerFilterOpt := func(bapp *baseapp.BaseApp) {
+	idPeerFilterOpt := func(bapp *BaseApp) {
 		bapp.SetIDPeerFilter(func(id string) abci.ResponseQuery {
 			require.Equal(t, "testid", id)
 			return abci.ResponseQuery{Code: uint32(4)}
@@ -1777,16 +974,16 @@ func TestGetMaximumBlockGas(t *testing.T) {
 	ctx := app.NewContext(true, tmproto.Header{})
 
 	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: 0}})
-	require.Equal(t, uint64(0), app.GetMaximumBlockGas(ctx))
+	require.Equal(t, uint64(0), app.getMaximumBlockGas(ctx))
 
 	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: -1}})
-	require.Equal(t, uint64(0), app.GetMaximumBlockGas(ctx))
+	require.Equal(t, uint64(0), app.getMaximumBlockGas(ctx))
 
 	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: 5000000}})
-	require.Equal(t, uint64(5000000), app.GetMaximumBlockGas(ctx))
+	require.Equal(t, uint64(5000000), app.getMaximumBlockGas(ctx))
 
 	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: -5000000}})
-	require.Panics(t, func() { app.GetMaximumBlockGas(ctx) })
+	require.Panics(t, func() { app.getMaximumBlockGas(ctx) })
 }
 
 func TestListSnapshots(t *testing.T) {
@@ -1968,47 +1165,6 @@ func (rtr *testCustomRouter) Route(ctx sdk.Context, path string) sdk.Handler {
 	return nil
 }
 
-func TestWithRouter(t *testing.T) {
-	// test increments in the handler
-	deliverKey := []byte("deliver-key")
-
-	txHandlerOpt := func(bapp *baseapp.BaseApp) {
-		customRouter := &testCustomRouter{routes: sync.Map{}}
-		r := sdk.NewRoute(routeMsgCounter, handlerMsgCounter(t, capKey1, deliverKey))
-		customRouter.AddRoute(r)
-		txHandler := middleware.NewRunMsgsTxHandler(middleware.NewMsgServiceRouter(interfaceRegistry), customRouter)
-		bapp.SetTxHandler(txHandler)
-	}
-	app := setupBaseApp(t, txHandlerOpt)
-	app.InitChain(abci.RequestInitChain{})
-
-	// Create same codec used in txDecoder
-	codec := codec.NewLegacyAmino()
-	registerTestCodec(codec)
-
-	nBlocks := 3
-	txPerHeight := 5
-
-	for blockN := 0; blockN < nBlocks; blockN++ {
-		header := tmproto.Header{Height: int64(blockN) + 1}
-		app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-		for i := 0; i < txPerHeight; i++ {
-			counter := int64(blockN*txPerHeight + i)
-			tx := newTxCounter(counter, counter)
-
-			txBytes, err := codec.Marshal(tx)
-			require.NoError(t, err)
-
-			res := app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-			require.True(t, res.IsOK(), fmt.Sprintf("%v", res))
-		}
-
-		app.EndBlock(abci.RequestEndBlock{})
-		app.Commit()
-	}
-}
-
 func TestBaseApp_EndBlock(t *testing.T) {
 	db := dbm.NewMemDB()
 	name := t.Name()
@@ -2020,7 +1176,7 @@ func TestBaseApp_EndBlock(t *testing.T) {
 		},
 	}
 
-	app := baseapp.NewBaseApp(name, logger, db, nil)
+	app := NewBaseApp(name, logger, db, nil)
 	app.SetParamStore(&paramStore{db: dbm.NewMemDB()})
 	app.InitChain(abci.RequestInitChain{
 		ConsensusParams: cp,
